requirements.txt
requests
python-dotenv
pyyaml
psycopg2-binary